% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocPoly.r
\name{allocPoly}
\alias{allocPoly}
\title{allocPoly}
\usage{
allocPoly(
  poly.lst,
  bounding.poly,
  ntows,
  mindist = 1,
  pool.size = 4,
  repeated.tows = NULL,
  map = NULL,
  lplace = "bottomleft",
  show.pool = F,
  UTMzone
)
}
\arguments{
\item{poly.lst}{= list containing PBSmapping::PolySet and PBSmapping::PolyData that describe strata polygons, PolyData may contain allocation to specifiy the number of station to be selected for each strata and repeats to indicate how many stations should be selected from repeated.tows}

\item{bounding.poly}{= PBSmapping::PolySet describing the area to be surveyed}

\item{ntows}{= total number of stations to be selected}

\item{mindist}{= minimum distance (km) or buffer between stations}

\item{pool.size}{= multiplyer to n.tows. This function first generates a pool of random locations, then randomly selects from the pool stations that fulfill the criteria (allocation by strata, buffer distance, etc.)}

\item{repeated.tows}{= PBSmapping::EventData of previous stations to be selected from if repeats exist in poly.lst[[2]] (PolyData)}

\item{map}{= preset location to pass to bioMap}

\item{lplace}{= placement of legend if !is.null(map)}

\item{show.pool}{= logical, if TRUE the initial pool of random stations is plotted on the map}

\item{UTMzone}{= PBSmapping::PolySet attribute, auto-generated if missing}
}
\description{
ALLOCATION by STRATIFIED RANDOM DESIGN based on bottom type (or another polygon based non-continuous stratifying variable)
}
\author{
Brad Hubley
}
