% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_desc.r
\docType{data}
\name{bathyPoly4}
\alias{bathyPoly4}
\title{This is the description of the bathyPoly4 data}
\description{
This is the description of the bathyPoly4 data
}
\references{
\url{data_blah.com}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\keyword{data}
