% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioMap.r
\name{bioMap}
\alias{bioMap}
\title{bioMap}
\usage{
bioMap(area = "custom", ylim = c(40, 52), xlim = c(-74, -47),
  mapRes = "HR", land.col = "wheat", title = "", nafo = NULL,
  boundaries = "LFAs", isobaths = seq(100, 1000, 100), bathcol = rgb(0, 0,
  1, 0.1), points.lst = NULL, pt.cex = 1, lines.lst = NULL,
  poly.lst = NULL, contours = NULL, image.lst = NULL,
  color.fun = tim.colors, zlim, grid = NULL, stippling = F, lol = F,
  labels = "lfa", labcex = 1.5, LT = T, plot.rivers = T,
  addsubareas = F, subsetSurveyStrata = NULL, basemap = F, ...)
}
\arguments{
\item{area}{= 'custom' where xlim & ylim are specified or select from area list below}

\item{mapRes}{= coastline detail ('LR' = low resolution, 'MR' = medium resolution, 'HR' = high resolution, 'UR' = ultra resolution)}

\item{title}{= plot title}

\item{boundaries}{= for ploting specific management boundaries ("lobster", "scallop", "snowcrab", "SummerSurveyStrata", "GeorgesSurveyStrata", "AmericanSurveyStrata")}

\item{bathcol}{= isobath line color, default is transparent blue}

\item{points.lst}{= points to overlay on map in PBSmapping format - list with 2 elements: 1st element is eventSet (EID, POS, X, Y), 2nd element is eventData (EID, pch, col, etc.)}

\item{lines.lst}{= lines to overlay on map in PBSmapping format - list with 2 elements: 1st element is polySet (PID, SID, POS, X, Y), 2nd element is polyData (PID, SID, lty, col, etc.)}

\item{poly.lst}{= polygons to overlay on map in PBSmapping format - list with 2 elements: 1st element is polySet (PID, SID, POS, X, Y), 2nd element is polyData (PID, SID, border, col, etc.)}

\item{contours}{= plots overlaping polygons as contours (same format as poly.lst)}

\item{image.lst}{= image to overlay on map - list with 3 elements (x, y, z), 'bathymetry' produces image from bathymetry data}

\item{color.fun}{= color function for image}

\item{zlim}{= zlim for image}

\item{grid}{= size of grid in degrees, default is no grid}

\item{stippling}{= adds stippling to land (purely for visual effect)}

\item{lol}{= adds water colored border to coastline (purely for visual effect)}

\item{isobath}{= plots bathymetry lines for specified depths from topex data}
}
\description{
a customizable mapping function that uses PBSmapping to plot fisheries data, bathymetry and mangement boundaries.
}
\examples{
bioMap(area='lfa34')
}
\author{
Brad Hubley
}
