% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contLegend.r
\name{contLegend}
\alias{contLegend}
\title{ContLegend}
\usage{
contLegend(..., lvls, Cont.data)
}
\arguments{
\item{...}{= undocumented}

\item{lvls}{= undocumented}

\item{Cont.data}{= undocumented}
}
\description{
Creates a legend for contour data
}
