% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contourGen.r
\name{contourGen}
\alias{contourGen}
\title{contourGen}
\usage{
contourGen(contour.dat, lvls, Poly, col = "YlGn", colorAdj)
}
\arguments{
\item{contour.dat}{= undocumented}

\item{lvls}{= undocumented}

\item{Poly}{= undocumented}

\item{col}{= undocumented}

\item{colorAdj}{= undocumented}
}
\description{
Generates contour polygons from image data
}
