% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.dd.dddd.r
\name{convert.dd.dddd}
\alias{convert.dd.dddd}
\title{convert.dd.dddd}
\usage{
convert.dd.dddd(x, format = "dec.deg")
}
\arguments{
\item{\code{x}}{= coordinates}

\item{\code{format}}{= dec.deg Acceptable values include dec.deg, deg.min}
}
\description{
This function can convert various formats to decimal degrees or to degrees-minutes
}
\examples{
convert.dd.dddd(4730.3)
[1] 47.505
convert.dd.dddd(47.505, 'deg.min')
[1] 4730.3
}
\author{
Brad Hubley
}
\seealso{
Other res: \code{\link{assignStation}}
}

