% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBathyLines.r
\name{createBathyLines}
\alias{createBathyLines}
\title{createBathyLines}
\usage{
createBathyLines(input.fp = file.path(project.datadirectory("bio.lobster"),
  "data", "maps", "topex"), input.fn = "topex.xyz", redo = T,
  interval = 10, batch = 100, range = c(10, 5000), save.by = 1000,
  digits = 3, dneg = T, output.fn = "bathyPoly")
}
