% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineStrata.r
\name{defineStrata}
\alias{defineStrata}
\title{defineStrata}
\usage{
defineStrata(char, nstrata = 4, min.str = 0, max.str, place = 0)
}
\description{
defines strata boundaries using the sqrt(f(y)) rule (Cochran 1977)
}

