% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designAnal.r
\name{designAnal}
\alias{designAnal}
\title{designAnal
description Analysis Efficiency Against Simple Random with "BIOsurvey" or "PEDstrata" package}
\usage{
designAnal(restrat.dat, restrat.areas, original.dat = NULL,
  original.areas = NULL, years = 2007:1981, plot = T, domain = T,
  biomass = F)
}
\description{
designAnal
description Analysis Efficiency Against Simple Random with "BIOsurvey" or "PEDstrata" package
}

