% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genran.r
\name{genran}
\alias{genran}
\title{description Generates random points in a polygon (PBSmapping format)}
\usage{
genran(npoints, bounding.poly, projection = "LL", mindist = NULL)
}
\description{
description Generates random points in a polygon (PBSmapping format)
}

