% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.r
\name{interpolation}
\alias{interpolation}
\title{interpolation}
\usage{
interpolation(contour.dat, ticks, nstrata, str.max, str.min, place = 0, aspr,
  interp.method = "gstat", res = 0.01, maxdist = Inf, nmax = 8,
  idp = 0.5, mod.type = "Sph", smooth = F, smooth.fun = median,
  smooth.procedure = 1, sres = 1/60.1, no.data = "0", blank = T,
  blank.dist, blank.eff = 0, blank.type = 2, log.dat = F,
  covariate.dat = NULL, subset.poly = NULL, regrid = F, subset.eff = NA,
  subscale = res)
}
\arguments{
\item{ticks}{= contour lines or strata definitions, 'define' will determine them according to the f(x) rule from Cochran 1977 (see Hubley et al. 2009)}

\item{nstrata}{= number of strata when ticks='define'}

\item{str.max}{= maximum value for stratifying variable (all values greater will be set at maximum)}

\item{str.min}{= minimum value for stratifying variable (all values lesser will be set to NA )}

\item{place}{= rounding place for defining strata}

\item{aspr}{= aspect ratio for for creating square grid, will determine from data if missing}

\item{interp.method}{= 'gstat' = gstat (idw) function from gstat library, 'krige' = krige function from gstat library, 'none' = no interpolation function}

\item{res}{= resolution for interpolation}

\item{maxdist, }{nmax, idp, mod.type = arguments to be passed to interpolation function (idp is inverse distance power)}

\item{smooth}{= logical, TRUE calls smoothing function}

\item{smooth.fun}{= applies smooth.fun to data over a grid}

\item{sres}{= resolution for smoothing (grid)}

\item{no.data}{= how to treat missing data when smoothing, default is to assume zero}

\item{blank}{= TRUE calls blanking function, included blanking distance, beyond which if no data are present zeros are assigned or blank.eff}

\item{blank.dist}{= blanking distance if missing will select the shortest distance to the most isolated point}

\item{blank.type}{= how spaced out the zeros are, 1 = avg. nearest neighbour distance, 2 = blanking distance}

\item{log.dat}{= logical, whether to log data}

\item{covariate.dat}{= covariate data typically used in kriging}

\item{subset.poly}{= inclusion polygon to subset spatially, 'square' used to close polygons (useful when not smoothing or blanking)}

\item{subset.eff}{= sets values to this outside the subset.poly}

\item{subscale}{= size of inset when subset.poly = 'square'}
}
\description{
preforms a spatial interpolation either inverse distance weighted or kriging.
}
\author{
Brad Hubley
}
