% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat2planar.r
\name{lonlat2planar}
\alias{lonlat2planar}
\title{lonlat2planar}
\usage{
lonlat2planar(
  x,
  proj.type,
  input_names = c("lon", "lat"),
  newnames = c("plon", "plat")
)
}
\arguments{
\item{x}{= undocumented}

\item{proj.type}{= undocumented}

\item{input_names}{= undocumented}

\item{newnames}{= undocumented}
}
\description{
convert lon/lat to a projected surface using proj
proj.type can be an internal code such as "utm20" or a proj4 argument
output scale is defined in the +units=km (default for bio) or +units=m (default for proj)
}
