% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.projection.params.r
\name{lookup.projection.params}
\alias{lookup.projection.params}
\title{lookup.projection.params}
\usage{
lookup.projection.params(x)
}
\arguments{
\item{x}{= undocumented}
}
\description{
Looks up projection parameters; see
\url{http://www.progonos.com/furuti/MapProj/Normal/TOC/cartTOC.html}
}
