% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planar2lonlat.r
\name{planar2lonlat}
\alias{planar2lonlat}
\title{planar2lonlat}
\usage{
planar2lonlat(x, proj.type, planar.coord.scale = NULL, newnames = c("lon",
  "lat"))
}
\arguments{
\item{x}{= undocumented}

\item{proj.type}{= undocumented}

\item{planar.coord.scale}{= undocumented}

\item{newnames}{= undocumented}
}
\description{
inverse projection: planar to lon/lat using projection. convert
planar coord systems (length scale defined in the CRS as +units=km or m)  to
lon-lat
}
