% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planarMap.r
\name{planarMap}
\alias{planarMap}
\title{planarMap}
\usage{
planarMap(
  xyz,
  depths = T,
  pts = NULL,
  colpts = F,
  annot = NULL,
  annot.cex = 2.2,
  scalebar = T,
  projection = "utm20",
  col.regions = T,
  datascale = seq(0, 1, l = 50),
  at = datascale,
  fn = paste("map", trunc(runif(1) * 1e+08), sep = ""),
  loc = tempdir(),
  corners = NULL,
  rez = c(1, 1),
  save = F,
  pt.cex = 0.5,
  pt.pch = 16,
  pt.col = "black",
  colorkey = NULL,
  fill = T,
  log.variable = F,
  interpolation = F,
  add.zeros = F,
  theta = 50,
  rev = F,
  ...
)
}
\arguments{
\item{xyz}{data to be plotted}

\item{depths}{= undocumented}

\item{pts}{data points to be overlayed on the map}

\item{colpts}{= undocumented}

\item{annot}{= undocumented}

\item{annot.cex}{= undocumented}

\item{scalebar}{= undocumented}

\item{projection}{to be used, see: lookup.projection.params}

\item{col.regions}{= undocumented}

\item{datascale}{for legend that describes the z variable}

\item{at}{= undocumented}

\item{fn}{= undocumented}

\item{loc}{= undocumented}

\item{corners}{2x2 dataframe that specify the corners of the map i.e. corners = data.frame(lon=c(-67.54,-56.5),lat=c(41,47.2))}

\item{rez}{= undocumented}

\item{save}{logical TRUE saves the map as a .png, FALSE prints to screen}

\item{pt.cex}{= undocumented}

\item{pt.pch}{= undocumented}

\item{pt.col}{= undocumented}

\item{colorkey}{= undocumented}

\item{fill}{= undocumented}

\item{log.variable}{= undocumented}

\item{interpolation}{logical, does thin plate spline interpolation of the z variable and plots the result}

\item{add.zeros}{uses zeroInflate() to add zeros where there is no data before the interpolation (assumes zero outside the sampled area)}

\item{theta}{interpolation parameter}

\item{rev}{= undocumented}

\item{...}{= undocumented}
}
\description{
a  mapping function that uses levelplot to plot data in planar coordinates
}
\author{
Brad Hubley (mostly copied from Jae Choi's bio.spacetime::map)
}
