% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reStratify.r
\name{reStratify}
\alias{reStratify}
\title{reStratify}
\usage{
reStratify(events, polys, lines = F)
}
\arguments{
\item{events}{= PBSmapping::EventData of sets}

\item{lines}{= logical, if TRUE X1, Y1, X2, Y2 must be supplied to describe the sets}

\item{strata}{= PBSmapping::PolyData of strata}
}
\description{
Assigns station number to tows based on proximity.
}
\author{
Brad Hubley
}
