% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.r
\name{smoothing}
\alias{smoothing}
\title{smoothing
description applies a function on spatial data over a grid}
\usage{
smoothing(dat, fun = mean, res = 0.01, aspr = 1.34564, no.data = "0",
  matrix = F, procedure = 1, subset.poly = NULL, expand = 0.1)
}
\description{
smoothing
description applies a function on spatial data over a grid
}

