% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blanking.r
\name{blanking}
\alias{blanking}
\title{blanking}
\usage{
blanking(surv.dat, blank.dist, aspr, type = 1, eff = 0, scale = 0.5,
  type.scaler = 0.5)
}
\arguments{
\item{surv.dat}{= data.frame of points containing X and Y coordinates}

\item{blank.dist}{= distance at which zeros are inserted into the data set, defaults to the largest nearest neighbour distance i.e. the distance to the most isolated point in the data}

\item{aspr}{= aspect ratio for for creating square grid}

\item{type}{= how spaced out the zeros are, 1 = avg. nearest neighbour distance, 2 = blanking distance}

\item{eff}{= what data is inserted, intended for zeros}

\item{scale}{= how far beyond the range of the data are points inserted}
}
\description{
incorporates blanking distance by including zeros spaced eqully at the average nearest nieghbour distance, default blanking distance is the nearest neighbour distance of the most isolated point. Called by interpolation.
}
\author{
Brad Hubley
}
