% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridData.r
\name{gridData}
\alias{gridData}
\title{gridData}
\usage{
gridData(Data, domain.poly, lvls, bcol = "YlGnBu", border = 1, FUN = mean,
  grid.size = 1, aspr = "calculate", sx, sy, ex, ey)
}
\arguments{
\item{Data}{= PBSmapping::EventData}

\item{domain.poly}{= polygon the defines the total area from which to construct the grid, dervied from data if missing}

\item{lvls}{= levels}

\item{bcol}{= RColorBrewer color palette}

\item{border}{= polygon border}

\item{FUN}{= summary function to apply to data within a grid cell}

\item{grid.size}{= size of the grid cells (km)}

\item{aspr}{= aspect ratio for for creating square grid, default is to calculate from data}

\item{sx}{= defines total extent to grid, defaults to the extent of the data}

\item{sy}{= defines total extent to grid, defaults to the extent of the data}

\item{ex}{= defines total extent to grid, defaults to the extent of the data}

\item{ey}{= defines total extent to grid, defaults to the extent of the data}
}
\description{
This function summarizes data onto a grid, creates PBSmapping polySet and polyData
}
\author{
Brad Hubley
}
