% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagePrep.r
\name{imagePrep}
\alias{imagePrep}
\title{imagePrep
description Called by interpolation to preform the interpolation and return image data}
\usage{
imagePrep(X, Y, Z, dat, aspr = 1.34564, res = 0.02, summary.dat = F,
  log.dat = T, method = "gstat", matrix.dat = T, idp = 0.5, nmax = 7,
  maxdist = Inf, linear = F, subset.poly = NULL, covariate.dat = NULL,
  regrid = F, mod.type = "Sph", subscale = 0.01)
}
\arguments{
\item{dat}{= a dataframe with 3 columns (longitude, latitude, variable to be mapped)}

\item{aspr}{= aspect ratio for a given latitude (default is for 45 deg.)}

\item{res}{= resolution of image in decimal degrees}
}
\description{
imagePrep
description Called by interpolation to preform the interpolation and return image data
}

