% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkTows.r
\name{linkTows}
\alias{linkTows}
\title{linkTows}
\usage{
linkTows(events, tows, expwin = 0.05, mindist = 10)
}
\arguments{
\item{events}{= PBSmapping::EventData of tows}

\item{tows}{=}

\item{expwin}{= expansion window beyond supplied points that defines the area examined}

\item{mindist}{=}
}
\description{
Assigns station number to tows based on proximity.
}
\author{
Brad Hubley
}
